+++ apps/feedahelp/src/components/ui/features/header/navbar/Navbar.tsx
@@ -3,7 +3,7 @@ import { Styled } from "./Navbar.styled";
 import MenuIcon from "@mui/icons-material/Menu";
 import Modal from "../../../../../../../../packages/ui/components/elements/Modal/GenericModal";
 import Image from "next/image";
-import LanguageSelections from "~/components/translationDropdown/LanguageSelections";
+import LanguageSelections from "~/components/ui/features/header/languageDropdown/LanguageSelections";
 import CurrencyDropDown from "../CurrencyDropDown/CurrencyDropDown";
